Working version of ctkCommandLineModuleWidget
This version uses a subclass of the FrontendQtGui in order
to construct the widgets.

Based on discussion with Sascha and Jc, we agree that some refactoring
of the classes is needed to better support using the xml-based
gui outside of the command line module framework.  Ideally the
ctkCmdLineModuleFrontendQtGui would be split into a part that
only depends on the moduleDescription and not on a moduleReference.
This should be done in the context of a re-work of the Frontend
system to be sure other specializations can be similarly refactored.